Add files via upload
Mean:
Definition: The mean, often referred to as the average, is calculated by summing up all values in a dataset and dividing by the total number of values.
Formula: 
Mean
=
Sum of all values
Number of values
Mean= 
Number of values
Sum of all values
​
 
Example: Consider the dataset [2, 4, 6, 8, 10]. The mean is calculated as 
2
+
4
+
6
+
8
+
10
5
=
6
5
2+4+6+8+10
​
 =6. In this case, the mean represents the central tendency of the data.
2. Median:
Definition: The median is the middle value of a dataset when it is arranged in ascending or descending order. If the dataset has an even number of values, the median is the average of the two middle values.
Example: For the dataset [3, 6, 8, 12, 15], the median is 8. If the dataset is [3, 6, 8, 12, 15, 18], the median is 
8
+
12
2
=
10
2
8+12
​
 =10. The median is less sensitive to extreme values than the mean.
3. Mode:
Definition: The mode is the value that appears most frequently in a dataset. A dataset can have one mode, more than one mode (multimodal), or no mode at all if all values are distinct.
Example: In the dataset [2, 4, 4, 6, 8, 8, 8, 10], the mode is 8 because it appears more frequently than any other value. If a dataset has no repeated values, it is said to have no mode (e.g., [1, 2, 3, 4, 5]).